neatogen dijkstra_sgd: return a 'size_t'
This removes some cast noise from prior changes.

+++ lib/neatogen/sgd.c
@@ -169,14 +169,14 @@ void sgd(graph_t *G, /* input graph */
   }
   term_sgd *terms = gv_calloc(n_terms, sizeof(term_sgd));
   // calculate term values through shortest paths
-  int offset = 0;
+  size_t offset = 0;
   graph_sgd *graph = extract_adjacency(G, model);
   for (size_t i = 0; i < n; i++) {
     if (!isFixed(GD_neato_nlist(G)[i])) {
       offset += dijkstra_sgd(graph, i, terms + offset);
     }
   }
-  assert((size_t)offset == n_terms);
+  assert(offset == n_terms);
   free_adjacency(graph);
   if (Verbose) {
     fprintf(stderr, " %.2f sec\n", elapsed_sec());
